% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGMAM.R
\name{calculateGeneticEffect}
\alias{calculateGeneticEffect}
\title{calculateGeneticEffect}
\usage{
calculateGeneticEffect(
  BECkey,
  cohortData,
  pixelGroupMap,
  transferTable,
  currentBEC,
  ecoregionMap,
  cohortDefinitionCols = c("pixelGroup", "speciesCode", "age", "Provenance")
)
}
\arguments{
\item{BECkey}{a key that matches BECraster code with transfer table}

\item{cohortData}{The LandR cohortData object}

\item{pixelGroupMap}{the pixelGroupMap needed to match cohorts with raster values}

\item{transferTable}{a table with genetic performance of species in each variant}

\item{currentBEC}{the current projected BEC zone}

\item{ecoregionMap}{a raster an RAT that matches ecoregionGroup to ecoregion}

\item{cohortDefinitionCols}{cohortData columns that determine individual cohorts}
}
\description{
Predict climate induced height reduction due to genetics
}
